requirements.txt
langgraph
langchain
langchain-core
langchain-openai
langsmith
openai
pydantic
python-dotenv
streamlit
pymongo
fastapi
uvicorn[standard]
python-multipart